refactor: Sample 도메인 수정
