+++ src/main/java/ex/sample/domain/sample/domain/SampleRepository.java
@@ -1,13 +1,12 @@
 package ex.sample.domain.sample.domain;
 
 import java.util.Optional;
-import java.util.UUID;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
 public interface SampleRepository {
 
-    Optional<Sample> findById(UUID id);
+    Optional<Sample> findById(Long id);
 
     Page<Sample> findAll(Pageable pageable);
