+++ src/main/java/ex/sample/domain/sample/dto/response/CreateSampleRes.java
@@ -1,9 +1,7 @@
 package ex.sample.domain.sample.dto.response;
 
-import java.util.UUID;
-
 public record CreateSampleRes(
-    UUID id,
+    Long id,
     String name,
     Long money
 ) {
