+++ src/main/java/ex/sample/domain/sample/dto/response/GetSampleRes.java
@@ -1,9 +1,7 @@
 package ex.sample.domain.sample.dto.response;
 
-import java.util.UUID;
-
 public record GetSampleRes(
-    UUID id,
+    Long id,
     String name,
     Long money
 ) {
