+++ src/main/java/ex/sample/domain/sample/implementation/SampleReader.java
@@ -5,11 +5,10 @@ import ex.sample.domain.sample.domain.SampleRepository;
 import ex.sample.global.annotation.Implementation;
 import ex.sample.global.exception.GlobalException;
 import ex.sample.global.response.ResponseCode;
-import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Slice;
 
 @Slf4j
 @Implementation
@@ -18,12 +17,12 @@ public class SampleReader {
 
     private final SampleRepository sampleRepository;
 
-    public Sample read(UUID id) {
+    public Sample read(Long id) {
         return sampleRepository.findById(id)
             .orElseThrow(() -> new GlobalException(ResponseCode.NOT_FOUND));
     }
 
-    public Slice<Sample> readAll(Pageable pageable) {
+    public Page<Sample> readAll(Pageable pageable) {
         return sampleRepository.findAll(pageable);
     }
 }
