+++ src/main/java/ex/sample/domain/sample/mapper/SampleMapper.java
@@ -1,14 +1,13 @@
 package ex.sample.domain.sample.mapper;
 
-import static org.mapstruct.MappingConstants.ComponentModel.SPRING;
-
 import ex.sample.domain.sample.domain.Sample;
 import ex.sample.domain.sample.dto.response.CreateSampleRes;
 import ex.sample.domain.sample.dto.response.GetSampleRes;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
+import org.mapstruct.MappingConstants.ComponentModel;
 
-@Mapper(componentModel = SPRING)
+@Mapper(componentModel = ComponentModel.SPRING)
 public interface SampleMapper {
 
     @Mapping(target = "money", source = "money.amount")
