+++ src/main/java/ex/sample/domain/sample/repository/SampleInMemoryRepository.java
@@ -1,4 +1,4 @@
-package ex.sample.domain.sample.infrastructure;
+package ex.sample.domain.sample.repository;
 
 import ex.sample.domain.sample.domain.Sample;
 import ex.sample.domain.sample.domain.SampleRepository;
@@ -6,7 +6,6 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
-import java.util.UUID;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
@@ -15,10 +14,10 @@ import org.springframework.stereotype.Repository;
 @Repository
 public class SampleInMemoryRepository implements SampleRepository {
 
-    private final Map<UUID, Sample> sampleMap = new HashMap<>();
+    private final Map<Long, Sample> sampleMap = new HashMap<>();
 
     @Override
-    public Optional<Sample> findById(UUID id) {
+    public Optional<Sample> findById(Long id) {
         return Optional.ofNullable(sampleMap.get(id));
     }
