+++ src/main/java/ex/sample/domain/sample/repository/SampleJpaRepository.java
@@ -1,12 +1,11 @@
-package ex.sample.domain.sample.infrastructure;
+package ex.sample.domain.sample.repository;
 
 import ex.sample.domain.sample.domain.Sample;
 import ex.sample.domain.sample.domain.SampleRepository;
-import java.util.UUID;
 import org.springframework.context.annotation.Primary;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 @Primary
-public interface SampleJpaRepository extends SampleRepository, JpaRepository<Sample, UUID> {
+public interface SampleJpaRepository extends SampleRepository, JpaRepository<Sample, Long> {
 
 }
