+++ src/main/java/ex/sample/domain/sample/service/SampleQueryService.java
@@ -1,14 +1,14 @@
-package ex.sample.domain.sample.application;
+package ex.sample.domain.sample.service;
 
 import ex.sample.domain.sample.domain.Sample;
 import ex.sample.domain.sample.dto.response.GetSampleRes;
 import ex.sample.domain.sample.implementation.SampleReader;
 import ex.sample.domain.sample.mapper.SampleMapper;
-import java.util.UUID;
+import ex.sample.global.response.CommonPageRes;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Slice;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -19,23 +19,24 @@ public class SampleQueryService {
 
     private final SampleReader sampleReader;
     private final SampleMapper sampleMapper;
-    
+
     /**
      * 샘플 단건 조회
      */
     @Transactional(readOnly = true)
-    public GetSampleRes getSample(UUID id) {
+    public GetSampleRes getSample(Long id) {
         Sample sample = sampleReader.read(id);
         return sampleMapper.toGetSampleRes(sample);
     }
 
     /**
      * 샘플 리스트 조회
-     * TODO : 페이징 처리를 List, Slice, Page 중 어떤 것을 사용할지 고민
      */
     @Transactional(readOnly = true)
-    public Slice<GetSampleRes> getSampleList(Pageable pageable) {
-        return sampleReader.readAll(pageable)
+    public CommonPageRes<GetSampleRes> getSampleList(Pageable pageable) {
+        Page<GetSampleRes> pages = sampleReader.readAll(pageable)
             .map(sampleMapper::toGetSampleRes);
+
+        return CommonPageRes.from(pages);
     }
 }
